requirements.txt
torch
ollama
mysql-connector-python
langchain_community
langchain_core
streamlit
sqlalchemy
transformers
sounddevice
numpy
python-dotenv